requirements.txt
Adafruit-ADS1x15==1.0.2
Adafruit-GPIO==1.0.3
Adafruit-PureIO==1.1.8
aiofiles==0.6.0
blinker==1.4
click==7.1.2
colorzero==1.1
gpiozero==1.5.1
h11==0.12.0
h2==4.0.0
hpack==4.0.0
Hypercorn==0.11.2
hyperframe==6.0.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
priority==1.3.0
pyserial==3.5
pyserial-asyncio==0.5
pyudev==0.22.0
Quart==0.14.1
Quart-CORS==0.4.0
RPi.GPIO==0.7.0
rshell==0.0.30
simple-pid==0.2.4
six==1.15.0
spidev==3.5
toml==0.10.2
watchdogdev==1.0.0
Werkzeug==1.0.1
wsproto==1.0.0
